moved routers to modules after review
