+++ src/rate/presentation/routers/exchanger.router.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import * as ExchangerController from '../../rate/presentation/controllers/exchanger.controller';
+import * as ExchangerController from '../controllers/exchanger.controller';
 
 export const exchangerRouter = Router();
 exchangerRouter.get('/', ExchangerController.getCurrentRate);
