+++ src/subscription/presentation/routers/subscription.router.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import * as SubscriptionController from '../controllers/subscription.controller';
+
+import { createSubscriprionRules } from '../rules/subscription.rules';
+
+export const subscriptionRouter = Router();
+
+subscriptionRouter.post('/', createSubscriprionRules, SubscriptionController.subscribe);
